particleに影響を与える場のテスト

+++ project/engine/Object/Particle.h
@@ -5,6 +5,6 @@
 struct Particle {
     Transform transform;
     Vector3 velocity;
-    float lifetime;
+    int lifetime;
     bool alive = false;
 };
